+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add an expanded beautify gradient picker with recent-preset promotion
+
 ## [1.6.18] - 2026-07-29
 
 ### Added